+++ cli/lib/tasks/verify.js
@@ -101,12 +101,8 @@ const runSmokeTest = (binaryDir, options) => {
     debug('smoke test command:', smokeTestCommand)
     debug('smoke test timeout %d ms', options.smokeTestTimeout)
 
-    const env = _.extend({}, process.env, {
-      ELECTRON_ENABLE_LOGGING: true,
-    })
-
     const stdioOptions = _.extend({}, {
-      env,
+      env: process.env,
       timeout: options.smokeTestTimeout,
     })