+++ cli/CHANGELOG.md
@@ -1,7 +1,7 @@
 <!-- See the ../guides/writing-the-cypress-changelog.md for details on writing the changelog. -->
 ## 13.6.0
 
-_Released 11/21/2023 (PENDING)_
+_Released 11/21/2023_
 
 **Features:**