+++ cli/CHANGELOG.md
@@ -1,4 +1,12 @@
 <!-- See the ../guides/writing-the-cypress-changelog.md for details on writing the changelog. -->
+## 13.6.1
+
+_Released 12/05/2023 (PENDING)_
+
+**Bugfixes:**
+
+- Cypress will now correctly log errors and debug logs on Linux machines. Fixes [#5051](https://github.com/cypress-io/cypress/issue/5051) and [#24713](https://github.com/cypress-io/cypress/issue/24713).
+
 ## 13.6.0
 
 _Released 11/21/2023 (PENDING)_